Reorganize course similarity code
